+++ gapoTT.cpp
@@ -895,7 +895,7 @@ void TF_gapoTT::OnClick(long Tag,unsigned long Col,unsigned long Row)
 			// podívám se, zda pohon, který je na øádku, kde došlo ke kliku má více objektù v KK režimu, pokud ano, musím projít všechny
 			 if(objekty[Row].pohon!=NULL)
 			 {
-					int pohon_n=objekty[Row].pohon->n;
+					unsigned int pohon_n=objekty[Row].pohon->n;
           indikator_skupin[pohon_n]=1;
 					if(F->d.v.vrat_pocet_objektu_vyuzivajici_pohon(objekty[Row].pohon->n,1) > 1)
 					{
@@ -940,7 +940,7 @@ void TF_gapoTT::OnClick(long Tag,unsigned long Col,unsigned long Row)
 		// podívám se, zda pohon, který je na øádku, kde došlo ke kliku má více objektù v KK režimu, pokud ano, musím projít všechny
 				if(objekty[Row].pohon!=NULL)
 				{
-					int pohon_n=objekty[Row].pohon->n;
+					unsigned int pohon_n=objekty[Row].pohon->n;
           indikator_skupin[pohon_n]=2;
 					if(F->d.v.vrat_pocet_objektu_vyuzivajici_pohon(objekty[Row].pohon->n,1) > 1)
 					{
@@ -995,7 +995,7 @@ void TF_gapoTT::OnClick(long Tag,unsigned long Col,unsigned long Row)
 		// podívám se, zda pohon, který je na øádku, kde došlo ke kliku má více objektù v KK režimu, pokud ano, musím projít všechny
 				if(objekty[Row].pohon!=NULL)
 				{
-					int pohon_n=objekty[Row].pohon->n;
+					unsigned int pohon_n=objekty[Row].pohon->n;
           indikator_skupin[pohon_n]=3;
 					if(F->d.v.vrat_pocet_objektu_vyuzivajici_pohon(objekty[Row].pohon->n,1) > 1)
 					{
@@ -1086,7 +1086,7 @@ void TF_gapoTT::OnClick(long Tag,unsigned long Col,unsigned long Row)
              {
 							 if(objekty[i].pohon!=NULL)
 							 {
-								 if(pohon_n==objekty[i].pohon->n)
+								 if((unsigned)pohon_n==objekty[i].pohon->n)
 								 {
 									 slouceny_radek=i;
 								 } //slouží k odlišení v calculate pro výpoèet, abych vìdìl na který slouèený øádek mám vrátit data
